changes on vuex and modulised userStore
